Rework and expand the nvidia role
Drivers are now installed using the official nvidia role as a dependency.
The driver installation is tested by executing nvidia-smi.

In addition to drivers, the core installation now includes the cuda toolkit
and the docker container runtime for gpus.

After the core components, libcudnn is installed. The libcudnn installation is
tested by installing, compiling and executing a sample project.

The sample project package is installed by sending a local copy to remote hosts.
Downloading it directly is problematic as it requires an NVIDIA Developer
Program Membership.