Add roles for GPU nodes currently in use by groups TIO and SHIP. (#17)
* Make the role "proxies" compatible with zsh.

* Add a role for installing seafile.

* Install miscellaneous development tools on g-nodes

* Add configuration files for regular docker users

The json configuration file for Docker is now added to all users that
are members of the docker group.

* Skip gracefully when no docker users exist.

* Do not restart docker when adding the json file

It's not necessary to restart the docker deamon when the
.docker/config.json file is added.

* Split the docker task file into two sets of tasks.

* Populate the docker group.

* Set up an NFS share for the tio group

These roles replicate an existing setup on GPU nodes in use by the TIO
group. Eventually, they will be superseded by the new file server.

* Add a play for the tio group.

* Fix a missing directory in the seafile role.

* Add a role replicating tio+ship user setup

The tio-users role recreates the current user setup on the nodes
in the tio group, which are not yet enrolled in the IPA domain. The list
of users is symlinked to the IPA user list to avoid duplication.

This role will become obsolete when the tio nodes are enrolled
in the IPA domain.

* Set a custom default shell for user vparmar.