Stop using the libcudnn8-samples package
The libcudnn8-samples package includes sample projects to verify
that libcudnn8 works correctly, but downloading the package requires
an Nvidia account, so it can't be distributed easily.

Since it's more hassle than it's worth, the verification step is simply
removed. As a quick sanity check, 'nvcc -V' is now executed instead.