Feature/resolve linting problems (#30)
The repository has been reorganized with a dedicated ansible directory which enables us to add other forms of IaC at a later stage as well (Terraform for example). Furthermore we can clearly distinguish ansible files from the rest. This way linting becomes easier.

All linting problems apart from one are solved. The one error that remains is very strange. It says that the linter does not recognize the module 'parted' in the c-nodes role. However when the linter is executed on the role directly not issues are found, only when site.yml is checked.