Fix issue #10
K8s plays are now consistently executed in the intended order. The RKE2
installation is first performed on the designed primary server, then on
the additional servers and finally on all agent nodes.

To make the setup easier to understand at a glance, the inventory group
with rke servers now has two sub-groups: one for the primary server and
one for the other servers.